% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBPair.R
\name{analyzeClinicalMeta}
\alias{analyzeClinicalMeta}
\title{Analyze clinical meta-analysis across datasets}
\usage{
analyzeClinicalMeta(patient_data_list)
}
\arguments{
\item{patient_data_list}{List of patient data from analyzeClinicalDrugResponse}
}
\value{
Meta-analysis results object or NULL if analysis couldn't be performed
}
\description{
Performs meta-analysis on clinical drug response data across multiple datasets
}
