% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncDrugOmicPair.R
\name{analyzeContinuousDrugOmic}
\alias{analyzeContinuousDrugOmic}
\title{Analyze continuous drug-omics pairs}
\usage{
analyzeContinuousDrugOmic(myPairs)
}
\arguments{
\item{myPairs}{List of paired omics-drug datasets from pairDrugOmic}
}
\value{
Meta-analysis results object or NULL if analysis couldn't be performed
}
\description{
Performs meta-analysis on continuous drug-omic pairs using Spearman correlation
}
