% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{analyzeDrugAWithSignature}
\alias{analyzeDrugAWithSignature}
\title{Analyze Drug A with B drug signature}
\usage{
analyzeDrugAWithSignature(
  select_drugs,
  connection,
  signature_genes,
  data_type,
  tumor_type
)
}
\arguments{
\item{select_drugs}{Name of drug A}

\item{connection}{Database connection}

\item{signature_genes}{B drug signature genes}

\item{data_type}{Filter by data type}

\item{tumor_type}{Filter by tumor type}
}
\value{
List containing patient data for Drug A
}
\description{
Loads and processes Drug A data for correlation analysis
}
\keyword{internal}
