% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{analyzeDrugBForSignature}
\alias{analyzeDrugBForSignature}
\title{Analyze Drug B to identify response signature}
\usage{
analyzeDrugBForSignature(
  drug_name,
  connection,
  top_n_genes,
  data_type,
  tumor_type,
  min_response_size,
  min_non_response_size
)
}
\arguments{
\item{drug_name}{Name of the drug}

\item{connection}{Database connection}

\item{top_n_genes}{Number of top genes to select}

\item{data_type}{Filter by data type}

\item{tumor_type}{Filter by tumor type}

\item{min_response_size}{Minimum response samples}

\item{min_non_response_size}{Minimum non-response samples}
}
\value{
List containing patient data and top genes per patient
}
\description{
Analyzes Drug B response data to identify upregulated genes in responders
}
\keyword{internal}
