% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{analyzeDrugBScores}
\alias{analyzeDrugBScores}
\title{Analyze B drug scores in NR vs Response groups}
\usage{
analyzeDrugBScores(drug_b_scores, drug_name)
}
\arguments{
\item{drug_b_scores}{List of B drug scores}

\item{drug_name}{Name of drug B}
}
\value{
List containing analysis results and plots
}
\description{
Compares B drug response scores between non-responders and responders
}
\keyword{internal}
