% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStratifiedAnalysis.R
\name{analyzeStratifiedData}
\alias{analyzeStratifiedData}
\title{Analyze stratified data}
\usage{
analyzeStratifiedData(
  omics_data,
  drug_data,
  feature_type,
  merged_enabled = TRUE,
  meta_enabled = TRUE
)
}
\arguments{
\item{omics_data}{Filtered omics data}

\item{drug_data}{Filtered drug data}

\item{feature_type}{Type of omics data}

\item{merged_enabled}{Whether to create merged dataset}

\item{meta_enabled}{Whether to perform meta-analysis}
}
\value{
Analysis results
}
\description{
Performs drug-omics analysis on pre-filtered stratified data
}
