% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{calculateDrugAScores}
\alias{calculateDrugAScores}
\title{Calculate A drug response scores using B drug signature}
\usage{
calculateDrugAScores(patient_data_list, signature_genes)
}
\arguments{
\item{patient_data_list}{List of patient expression data from Drug A}

\item{signature_genes}{Final signature genes from Drug B analysis}
}
\value{
List of scores for each patient in Drug A
}
\description{
Calculates drug A response scores using the signature genes from drug B
}
\keyword{internal}
