% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{calculateDrugBScores}
\alias{calculateDrugBScores}
\title{Calculate B drug response scores using ssGSEA}
\usage{
calculateDrugBScores(patient_data_list)
}
\arguments{
\item{patient_data_list}{List of patient expression data with top_genes included}
}
\value{
List of scores for each patient
}
\description{
Calculates drug response scores using ssGSEA on each patient's top genes
}
\keyword{internal}
