% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDB.R
\name{createClinicalForestPlot}
\alias{createClinicalForestPlot}
\title{Create a forest plot for clinical meta-analysis results}
\usage{
createClinicalForestPlot(
  meta_obj,
  xlab = "Effect Size (95\% CI)",
  show_common = FALSE
)
}
\arguments{
\item{meta_obj}{Meta-analysis object from metagen() function}

\item{xlab}{Label for x-axis}

\item{show_common}{Logical, whether to show common effect model}
}
\value{
A forest plot visualization
}
\description{
Creates a standardized forest plot for visualizing clinical meta-analysis results
}
