% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStratifiedAnalysis.R
\name{createMultiDromaSetFromList}
\alias{createMultiDromaSetFromList}
\title{Create MultiDromaSet from list of DromaSet objects}
\usage{
createMultiDromaSetFromList(dromaset_list, group_name)
}
\arguments{
\item{dromaset_list}{List of DromaSet objects}

\item{group_name}{Name for the group}
}
\value{
MultiDromaSet object
}
\description{
Helper function to create a MultiDromaSet from a list of DromaSet objects
}
