% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{estimateTimeRemaining}
\alias{estimateTimeRemaining}
\title{Calculate estimated time remaining based on progress}
\usage{
estimateTimeRemaining(done, total, elapsed_time)
}
\arguments{
\item{done}{Number of items processed}

\item{total}{Total number of items}

\item{elapsed_time}{Time elapsed so far in seconds}
}
\value{
Estimated time remaining in seconds
}
\description{
Estimates remaining time for a batch process based on current progress
}
