% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStratifiedAnalysis.R
\name{filterStratifiedData}
\alias{filterStratifiedData}
\title{Filter drug data for stratified analysis}
\usage{
filterStratifiedData(loaded_data, stratification_info, stratum)
}
\arguments{
\item{loaded_data}{List containing loaded omics and drug data}

\item{stratification_info}{Stratification information from getStratificationInfo}

\item{stratum}{Which stratum ("sensitive" or "resistant")}
}
\value{
List with filtered drug data (omics data is returned unchanged)
}
\description{
Filters loaded drug data to keep only samples from specified stratum.
Only filters datasets that are present in stratification_info.
}
