% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairDataPairing.R
\name{formatTime}
\alias{formatTime}
\title{Format seconds into a human-readable time string}
\usage{
formatTime(seconds)
}
\arguments{
\item{seconds}{Number of seconds}
}
\value{
Formatted time string
}
\description{
Converts seconds into a more readable time format (hours, minutes, seconds)
}
