% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStatMultiDromaSet.R
\name{generateCharacteristicsPlot}
\alias{generateCharacteristicsPlot}
\title{Generate characteristics plot using database information}
\usage{
generateCharacteristicsPlot(project_info, reverse = FALSE)
}
\arguments{
\item{project_info}{Data frame with project information from database}

\item{reverse}{Logical, whether to reverse axes (projects on x-axis, features on y-axis). Default FALSE}
}
\value{
ggplot object showing data availability
}
\description{
Creates a dot plot showing which data types (including drug and molecular) are available in each project
}
