% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStatMultiDromaSet.R
\name{generateDrugMOAPlot}
\alias{generateDrugMOAPlot}
\title{Generate drug MOA treemap plot from database annotations}
\usage{
generateDrugMOAPlot(drug_annotations)
}
\arguments{
\item{drug_annotations}{Data frame from getDROMAAnnotation("drug") with drug metadata}
}
\value{
ggplot object of drug MOA treemap
}
\description{
Creates a treemap visualization of drug mechanisms of action using database annotations
}
