% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairBatchFeature.R
\name{getAndValidateFeatureList}
\alias{getAndValidateFeatureList}
\title{Get and validate feature list from DromaSet object}
\usage{
getAndValidateFeatureList(
  dromaset_object,
  feature_type,
  feature_names = NULL,
  data_type = "all",
  tumor_type = "all"
)
}
\arguments{
\item{dromaset_object}{Either a DromaSet or MultiDromaSet object}

\item{feature_type}{Type of features to retrieve}

\item{feature_names}{Optional vector of specific feature names to validate (default: NULL for all features)}

\item{data_type}{Filter by data type}

\item{tumor_type}{Filter by tumor type}
}
\value{
Character vector of valid feature names
}
\description{
Get and validate feature list from DromaSet object
}
\keyword{internal}
