% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncGSVA.R
\name{getGeneSetByName}
\alias{getGeneSetByName}
\title{Get gene set by name from predefined gene sets}
\usage{
getGeneSetByName(gene_set_name, gmt_file = NULL)
}
\arguments{
\item{gene_set_name}{Name of the gene set to retrieve}

\item{gmt_file}{Optional path to GMT file. If NULL, uses the default Hallmark gene sets file}
}
\value{
A character vector of gene symbols in the gene set
}
\description{
Retrieves a specific gene set from the predefined GMT file
}
\examples{
\dontrun{
# Get HALLMARK_APOPTOSIS gene set
apoptosis_genes <- getGeneSetByName("HALLMARK_APOPTOSIS")
}
}
