% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{getSampleMetadata}
\alias{getSampleMetadata}
\title{Get sample metadata for discrete vs discrete analysis}
\usage{
getSampleMetadata(dromaset_object, feature1_type, feature2_type)
}
\arguments{
\item{dromaset_object}{Either a DromaSet or MultiDromaSet object}

\item{feature1_type}{Type of first feature}

\item{feature2_type}{Type of second feature}
}
\value{
Data frame with sample metadata
}
\description{
Get sample metadata for discrete vs discrete analysis
}
\keyword{internal}
