% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncDataLoaders.R
\name{isZscoreNormalized}
\alias{isZscoreNormalized}
\title{Check if data has been z-score normalized}
\usage{
isZscoreNormalized(data)
}
\arguments{
\item{data}{Matrix or data frame to check}
}
\value{
Logical indicating whether data has been z-score normalized
}
\description{
Utility function to check if data has the z-score normalization attribute
}
\examples{
\dontrun{
# Check if data is normalized
is_normalized <- isZscoreNormalized(my_data)
}
}
