% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncGSVA.R
\name{loadGeneSetsFromGMT}
\alias{loadGeneSetsFromGMT}
\title{Load gene sets from GMT file}
\usage{
loadGeneSetsFromGMT(gmt_file)
}
\arguments{
\item{gmt_file}{Path to the GMT file}
}
\value{
A named list where each element is a character vector of gene symbols for a gene set
}
\description{
Reads a GMT (Gene Matrix Transposed) file and returns a list of gene sets
}
\examples{
\dontrun{
# Load Hallmark gene sets
gmt_path <- system.file("extdata", "h.all.v2025.1.Hs.symbols.gmt", package = "DROMA")
gene_sets <- loadGeneSetsFromGMT(gmt_path)
}
}
