% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{metaCalcConCon}
\alias{metaCalcConCon}
\title{Calculate meta-analysis for continuous vs continuous features}
\usage{
metaCalcConCon(selected_pair)
}
\arguments{
\item{selected_pair}{List of paired data}
}
\value{
Meta-analysis result object or NULL if insufficient data
}
\description{
Calculate meta-analysis for continuous vs continuous features
}
