% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{metaCalcConDis}
\alias{metaCalcConDis}
\title{Calculate meta-analysis for continuous vs discrete features}
\usage{
metaCalcConDis(selected_pair)
}
\arguments{
\item{selected_pair}{List of paired data}
}
\value{
Meta-analysis result object or NULL if insufficient data
}
\description{
Calculate meta-analysis for continuous vs discrete features
}
