% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{metaCalcDisDis}
\alias{metaCalcDisDis}
\title{Calculate meta-analysis for discrete vs discrete features}
\usage{
metaCalcDisDis(selected_pair)
}
\arguments{
\item{selected_pair}{List of paired data}
}
\value{
Meta-analysis result object or NULL if insufficient data
}
\description{
Calculate meta-analysis for discrete vs discrete features
}
