% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{pairContinuousFeatures}
\alias{pairContinuousFeatures}
\title{Pair continuous with continuous data}
\usage{
pairContinuousFeatures(myOmics, myDrugs)
}
\arguments{
\item{myOmics}{First feature data list}

\item{myDrugs}{Second feature data list}
}
\value{
List of paired data
}
\description{
Creates paired datasets of two continuous features
}
