% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{pairDiscreteDiscrete}
\alias{pairDiscreteDiscrete}
\title{Pair discrete with discrete data}
\usage{
pairDiscreteDiscrete(
  my_feas1,
  my_feas2,
  feature1_type,
  feature2_type,
  samples_search
)
}
\arguments{
\item{my_feas1}{First discrete feature data list (samples with feature present)}

\item{my_feas2}{Second discrete feature data list (samples with feature present)}

\item{feature1_type}{Type of first feature}

\item{feature2_type}{Type of second feature}

\item{samples_search}{Reference data for all cells}
}
\value{
List of paired data with contingency tables
}
\description{
Creates paired datasets of two discrete features with contingency tables
}
