% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairDrugOmicPair.R
\name{pairDiscreteDrugOmic}
\alias{pairDiscreteDrugOmic}
\title{Pair discrete omics and drug data}
\usage{
pairDiscreteDrugOmic(myOmics, myDrugs, merged = FALSE)
}
\arguments{
\item{myOmics}{List of discrete omics data (samples with feature present)}

\item{myDrugs}{List of drug response data vectors}

\item{merged}{Logical, if TRUE, creates an additional merged dataset combining all pairs}
}
\value{
A list of paired drug-omics datasets with yes/no groups
}
\description{
Creates paired datasets of discrete omics and drug response data
}
