% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairDrugOmicPair.R
\name{pairDrugOmic}
\alias{pairDrugOmic}
\title{Pair continuous omics and drug data}
\usage{
pairDrugOmic(myOmics, myDrugs, merged = FALSE)
}
\arguments{
\item{myOmics}{List of omics data vectors}

\item{myDrugs}{List of drug response data vectors}

\item{merged}{Logical, if TRUE, creates an additional merged dataset combining all pairs}
}
\value{
A list of paired omics-drug datasets
}
\description{
Creates paired datasets of omics and drug response data from multiple sources
}
