% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{performCorrelationMetaAnalysis}
\alias{performCorrelationMetaAnalysis}
\title{Perform meta-analysis for correlations using Fisher's z transformation}
\usage{
performCorrelationMetaAnalysis(correlations, sample_sizes, study_labels)
}
\arguments{
\item{correlations}{Vector of correlation coefficients}

\item{sample_sizes}{Vector of sample sizes}

\item{study_labels}{Vector of study labels}
}
\value{
Meta-analysis result from meta::metagen or NULL
}
\description{
Performs meta-analysis on correlation coefficients using Fisher's z transformation
}
\keyword{internal}
