% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{performMetaAnalysisForCorrelations}
\alias{performMetaAnalysisForCorrelations}
\title{Perform meta-analysis for correlation differences}
\usage{
performMetaAnalysisForCorrelations(correlation_data)
}
\arguments{
\item{correlation_data}{List of correlation data from each study}
}
\value{
Meta-analysis result
}
\description{
Performs meta-analysis on correlation differences between groups
}
\keyword{internal}
