% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDB.R
\name{plotAllClinicaldatasets}
\alias{plotAllClinicaldatasets}
\title{Plot clinical drug response for all datasets}
\usage{
plotAllClinicaldatasets(patient_data_list, select_omics, select_drugs)
}
\arguments{
\item{patient_data_list}{List of patient data}

\item{select_omics}{Name of the omics feature}

\item{select_drugs}{Name of the drug}
}
\value{
A combined plot with all patient comparisons or NULL if no valid data
}
\description{
Creates and combines plots for all clinical datasets showing drug response differences
}
