% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncDrugOmicPair.R
\name{plotAllContinuousDrugOmic}
\alias{plotAllContinuousDrugOmic}
\title{Plot all continuous drug-omic pairs}
\usage{
plotAllContinuousDrugOmic(pairs_list)
}
\arguments{
\item{pairs_list}{List of paired drug-omic datasets}
}
\value{
A combined plot with all drug-omic correlations or NULL if no valid pairs
}
\description{
Creates and combines plots for all continuous drug-omic pairs
}
