% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairDrugOmicPair.R
\name{plotAllDiscreteDrugOmic}
\alias{plotAllDiscreteDrugOmic}
\title{Plot all discrete drug-omic pairs}
\usage{
plotAllDiscreteDrugOmic(pairs_list)
}
\arguments{
\item{pairs_list}{List of paired drug-omic datasets with yes/no groups}
}
\value{
A combined plot with all discrete drug-omic comparisons or NULL if no valid pairs
}
\description{
Creates and combines plots for all discrete drug-omic pairs
}
