% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncVisualization.R
\name{plotContinuousComparison}
\alias{plotContinuousComparison}
\title{Plot continuous comparison}
\usage{
plotContinuousComparison(
  data,
  cont_column,
  value_column = "value",
  value_label = "Drug Sensitivity",
  title = NULL
)
}
\arguments{
\item{data}{Data frame containing the variables to plot.}

\item{cont_column}{Name of the continuous column to use for x-axis.}

\item{value_column}{Name of the value column to use for y-axis.}

\item{value_label}{Label for the value variable.}

\item{title}{Plot title.}
}
\value{
A ggplot2 object with the comparison plot.
}
\description{
Creates a scatter plot showing the association between a
  continuous variable and a response value.
}
