% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairDrugOmicPair.R
\name{plotContinuousDrugOmic}
\alias{plotContinuousDrugOmic}
\title{Plot correlation between continuous drug and omic data}
\usage{
plotContinuousDrugOmic(omic_values, drug_values, study_name)
}
\arguments{
\item{omic_values}{Vector of omic feature values}

\item{drug_values}{Vector of drug response values}

\item{study_name}{Name of the study for plot title}
}
\value{
A ggplot2 object with scatter plot and correlation statistics
}
\description{
Creates a scatter plot with correlation statistics for a single drug-omic pair
}
