% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncDrugOmicPair.R
\name{plotDiscreteDrugOmic}
\alias{plotDiscreteDrugOmic}
\title{Plot drug response by discrete omic feature}
\usage{
plotDiscreteDrugOmic(yes_values, no_values, study_name)
}
\arguments{
\item{yes_values}{Drug response values for samples with the feature}

\item{no_values}{Drug response values for samples without the feature}

\item{study_name}{Name of the study for plot title}
}
\value{
A ggplot2 object with boxplot and statistical test
}
\description{
Creates a boxplot comparing drug response between samples with and without a discrete omic feature
}
