% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairBatchFeature.R
\name{processFeaturePair}
\alias{processFeaturePair}
\title{Process a single feature pair for batch analysis}
\usage{
processFeaturePair(
  feature_idx,
  feature2_list,
  dromaset_object,
  preloaded_feature2_data,
  selected_feas1,
  feature1_type,
  feature2_type,
  is_continuous1,
  is_continuous2,
  data_type,
  tumor_type,
  overlap_only,
  samples_search,
  sample_intersection_cache,
  progress_callback = NULL,
  start_time = NULL,
  verbose = TRUE
)
}
\arguments{
\item{feature_idx}{Index of feature2 in feature2_list}

\item{feature2_list}{Vector of feature2 names to test}

\item{dromaset_object}{DromaSet or MultiDromaSet object}

\item{preloaded_feature2_data}{Pre-loaded feature2 data matrix or dataframe}

\item{selected_feas1}{Feature1 data}

\item{feature1_type}{Type of feature1}

\item{feature2_type}{Type of feature2}

\item{is_continuous1}{Whether feature1 is continuous}

\item{is_continuous2}{Whether feature2 is continuous}

\item{data_type}{Filter by data type}

\item{tumor_type}{Filter by tumor type}

\item{overlap_only}{Whether to use only overlapping samples}

\item{samples_search}{Sample metadata for discrete vs discrete analysis}

\item{sample_intersection_cache}{Cached sample intersections}

\item{progress_callback}{Optional progress callback function}

\item{start_time}{Start time for progress tracking}

\item{verbose}{Logical, whether to display detailed messages}
}
\value{
Data frame with p_value, effect_size, n_datasets or NULL
}
\description{
Process a single feature pair for batch analysis
}
\keyword{internal}
