% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{selectFinalSignatureGenes}
\alias{selectFinalSignatureGenes}
\title{Select final signature genes based on intersection and logFC ranking}
\usage{
selectFinalSignatureGenes(top_genes_per_patient)
}
\arguments{
\item{top_genes_per_patient}{List of top genes for each patient}
}
\value{
Vector of final signature genes
}
\description{
Selects final signature genes by prioritizing genes that appear
in multiple datasets and have high logFC values
}
\keyword{internal}
