% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{analyzeCorrelationsInDrugA}
\alias{analyzeCorrelationsInDrugA}
\title{Analyze correlations in Drug A datasets}
\usage{
analyzeCorrelationsInDrugA(drug_a_data, select_omics, drug_b_scores)
}
\arguments{
\item{drug_a_data}{Drug A patient data}

\item{select_omics}{Name of the omics feature to analyze}

\item{drug_b_scores}{B drug scores for reference}
}
\value{
List containing correlation results and plots
}
\description{
Analyzes correlation between specific omics feature expression and B drug scores
}
\keyword{internal}
