% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncZscoreWhole.R
\name{applyZscoreNormalization}
\alias{applyZscoreNormalization}
\title{Apply Z-score normalization to all continuous data}
\usage{
applyZscoreNormalization()
}
\value{
None, directly modifies objects in the global environment
}
\description{
Applies Z-score normalization to all loaded omics and drug datasets in the global environment
}
