% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStratifiedAnalysis.R
\name{compareStratifiedResults}
\alias{compareStratifiedResults}
\title{Compare results between strata}
\usage{
compareStratifiedResults(
  sensitive_result,
  resistant_result,
  stratification_info
)
}
\arguments{
\item{sensitive_result}{Analysis result from sensitive group}

\item{resistant_result}{Analysis result from resistant group}

\item{stratification_info}{Information about stratification}
}
\value{
Comparison results with statistical tests
}
\description{
Compares association results between sensitive and resistant groups
}
