% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStratifiedAnalysis.R
\name{compareStratifiedResults}
\alias{compareStratifiedResults}
\title{Compare results between strata}
\usage{
compareStratifiedResults(
  sensitive_result,
  resistant_result,
  stratification_info,
  select_omics_type,
  select_omics,
  select_drugs
)
}
\arguments{
\item{sensitive_result}{Analysis result from sensitive group}

\item{resistant_result}{Analysis result from resistant group}

\item{stratification_info}{Information about stratification}

\item{select_omics_type}{Type of omics data analyzed}

\item{select_omics}{Name of the omics feature}

\item{select_drugs}{Name of the drug analyzed}
}
\value{
Comparison results with statistical tests
}
\description{
Compares association results between sensitive and resistant groups
}
