% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairVisualization.R
\name{createPlotWithCommonAxes}
\alias{createPlotWithCommonAxes}
\title{Create plot with common axis labels}
\usage{
createPlotWithCommonAxes(
  p,
  x_title = "Common X-Axis Title",
  y_title = "Common Y-Axis Title"
)
}
\arguments{
\item{p}{A patchwork object containing multiple plots}

\item{x_title}{Common x-axis title}

\item{y_title}{Common y-axis title}
}
\value{
A function that generates the plot when called
}
\description{
Creates a plot with common axis labels for multiple subplots
}
