% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncDrugFeature.R
\name{createSingleRankPlot}
\alias{createSingleRankPlot}
\title{Create a single rank plot (internal function)}
\usage{
createSingleRankPlot(
  drug_data,
  select_drugs,
  highlight,
  color,
  zscore,
  value_column,
  point_size,
  highlight_alpha,
  merge,
  projects = NULL
)
}
\arguments{
\item{drug_data}{Data frame with drug sensitivity data}

\item{select_drugs}{Name of the drug}

\item{highlight}{Highlighting specification}

\item{color}{Coloring specification}

\item{zscore}{Whether using zscore values}

\item{value_column}{Which value column to use}

\item{point_size}{Size of points}

\item{highlight_alpha}{Alpha for non-highlighted points}

\item{merge}{Whether this is a merged plot}

\item{projects}{Optional project name for title}
}
\value{
ggplot2 object
}
\description{
Create a single rank plot (internal function)
}
