% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStratifiedAnalysis.R
\name{createStratifiedComparisonPlot}
\alias{createStratifiedComparisonPlot}
\title{Create stratified comparison plot}
\usage{
createStratifiedComparisonPlot(
  sensitive_result,
  resistant_result,
  select_omics,
  select_drugs,
  select_omics_type,
  p_value_digits = 3,
  add_difference_row = TRUE
)
}
\arguments{
\item{sensitive_result}{Results from sensitive group}

\item{resistant_result}{Results from resistant group}

\item{select_omics}{Name of omics feature}

\item{select_drugs}{Name of drug}

\item{select_omics_type}{Type of omics data}

\item{p_value_digits}{Number of decimal places for p-values (default: 3)}

\item{add_difference_row}{Whether to add a row showing difference between groups (default: TRUE)}
}
\value{
ggplot object
}
\description{
Creates a visualization comparing results between strata
}
