% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStratifiedAnalysis.R
\name{createStratifiedComparisonPlot}
\alias{createStratifiedComparisonPlot}
\title{Create stratified comparison plot}
\usage{
createStratifiedComparisonPlot(
  sensitive_result,
  resistant_result,
  select_omics,
  select_drugs
)
}
\arguments{
\item{sensitive_result}{Results from sensitive group}

\item{resistant_result}{Results from resistant group}

\item{select_omics}{Name of omics feature}

\item{select_drugs}{Name of drug}
}
\value{
ggplot object
}
\description{
Creates a visualization comparing results between strata
}
