% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{filterFeatureData}
\alias{filterFeatureData}
\title{Filter feature data to ensure minimum sample size}
\usage{
filterFeatureData(feature_data, min_samples = 3)
}
\arguments{
\item{feature_data}{List of feature data by dataset}

\item{min_samples}{Minimum number of samples required per dataset}
}
\value{
Filtered feature data
}
\description{
Filter feature data to ensure minimum sample size
}
\keyword{internal}
