% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncStatMultiDromaSet.R
\name{generateCharacteristicsPlot}
\alias{generateCharacteristicsPlot}
\title{Generate characteristics plot using database information}
\usage{
generateCharacteristicsPlot(project_info)
}
\arguments{
\item{project_info}{Data frame with project information from database}
}
\value{
ggplot object showing data availability
}
\description{
Creates a dot plot showing which data types (including drug and molecular) are available in each project
}
