% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairDataLoading.R
\name{loadFeatureData}
\alias{loadFeatureData}
\title{Load feature data from DromaSet or MultiDromaSet}
\usage{
loadFeatureData(
  dromaset_object,
  feature_type,
  select_features,
  data_type = "all",
  tumor_type = "all",
  overlap_only = FALSE,
  is_continuous = TRUE,
  return_all_samples = FALSE,
  zscore = TRUE
)
}
\arguments{
\item{dromaset_object}{Either a DromaSet or MultiDromaSet object}

\item{feature_type}{Type of feature to load}

\item{select_features}{Name(s) of feature(s) to load. Can be a single feature name or a vector of feature names. If multiple features are provided for continuous data, returns a matrix with all features.}

\item{data_type}{Filter by data type}

\item{tumor_type}{Filter by tumor type}

\item{overlap_only}{For MultiDromaSet, whether to use only overlapping samples}

\item{is_continuous}{Whether the feature is continuous}

\item{return_all_samples}{For discrete features, whether to return all profiled samples in addition to feature-present samples (default: FALSE)}

\item{zscore}{Whether to apply z-score normalization (default: TRUE)}
}
\value{
For single feature (continuous): List of named vectors by dataset. For multiple features (continuous): List of matrices by dataset. For discrete features: if return_all_samples=FALSE, vector of sample IDs; if return_all_samples=TRUE, list with 'present' and 'all' components.
}
\description{
Load feature data from DromaSet or MultiDromaSet
}
\keyword{internal}
