% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairBatchFeature.R
\name{loadFeatureData}
\alias{loadFeatureData}
\title{Load feature data from DromaSet or MultiDromaSet}
\usage{
loadFeatureData(
  dromaset_object,
  feature_type,
  feature_name,
  data_type = "all",
  tumor_type = "all",
  overlap_only = FALSE,
  is_continuous = TRUE
)
}
\arguments{
\item{dromaset_object}{Either a DromaSet or MultiDromaSet object}

\item{feature_type}{Type of feature to load}

\item{feature_name}{Name of feature to load}

\item{data_type}{Filter by data type}

\item{tumor_type}{Filter by tumor type}

\item{overlap_only}{For MultiDromaSet, whether to use only overlapping samples}

\item{is_continuous}{Whether the feature is continuous}
}
\value{
List of feature data by dataset
}
\description{
Load feature data from DromaSet or MultiDromaSet
}
\keyword{internal}
