% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairMetaAnalysis.R
\name{metaCalcConDis}
\alias{metaCalcConDis}
\title{Meta-analysis for continuous vs discrete features}
\usage{
metaCalcConDis(selected_pair)
}
\arguments{
\item{selected_pair}{List of paired data with continuous and discrete variables}
}
\value{
Meta-analysis result object or NULL if insufficient data
}
\description{
Performs meta-analysis comparing continuous values between discrete groups
}
