% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairMetaAnalysis.R
\name{metaCalcDisDis}
\alias{metaCalcDisDis}
\title{Meta-analysis for discrete vs discrete features}
\usage{
metaCalcDisDis(selected_pair)
}
\arguments{
\item{selected_pair}{List of paired data with contingency tables}
}
\value{
Meta-analysis result object or NULL if insufficient data
}
\description{
Performs meta-analysis on two discrete features using odds ratio
}
