% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncPairDataPairing.R
\name{pairDiscreteFeatures}
\alias{pairDiscreteFeatures}
\title{Pair discrete with continuous data}
\usage{
pairDiscreteFeatures(discrete_dataset, continuous_dataset, merged = FALSE)
}
\arguments{
\item{discrete_dataset}{List with discrete feature data (samples with feature present)}

\item{continuous_dataset}{List with continuous feature data (numeric values)}

\item{merged}{Logical, if TRUE, creates an additional merged dataset}
}
\value{
List of paired data with yes/no groups for continuous values
}
\description{
Creates paired datasets of discrete and continuous features
}
