% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncBatchFeature.R
\name{pairDiscreteFeatures}
\alias{pairDiscreteFeatures}
\title{Pair discrete with continuous data}
\usage{
pairDiscreteFeatures(myOmics, myDrugs)
}
\arguments{
\item{myOmics}{Discrete feature data list (samples with feature present)}

\item{myDrugs}{Continuous feature data list (numeric values)}
}
\value{
List of paired data with yes/no groups
}
\description{
Creates paired datasets of discrete and continuous features
}
