% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncCTRDBStratifiedAnalysis.R
\name{performMetaAnalysisForCorrelations}
\alias{performMetaAnalysisForCorrelations}
\title{Perform meta-analysis for correlation differences}
\usage{
performMetaAnalysisForCorrelations(meta_data)
}
\arguments{
\item{meta_data}{Data frame with correlation results from all studies and groups}
}
\value{
Meta-analysis result
}
\description{
Performs meta-analysis on correlation differences between groups
}
\keyword{internal}
