% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncVisualization.R
\name{plotEnrichDotplot}
\alias{plotEnrichDotplot}
\title{Create a dotplot for enrichment results}
\usage{
plotEnrichDotplot(
  enrich_result,
  use_padj = TRUE,
  showCategory = 10,
  title = NULL,
  x_label = NULL,
  size_by = c("Count", "GeneRatio"),
  color_low = "#BEBADAFF",
  color_high = "#FB8072FF",
  ...
)
}
\arguments{
\item{enrich_result}{An enrichResult object from clusterProfiler.}

\item{use_padj}{Logical, whether to use adjusted p-value.}

\item{showCategory}{Number of top pathways to display.}

\item{title}{Plot title.}

\item{x_label}{Label for x-axis.}

\item{size_by}{Variable for point size.}

\item{color_low}{Low-significance color.}

\item{color_high}{High-significance color.}

\item{...}{Additional parameters passed to ggplot2 functions.}
}
\value{
A ggplot2 object with dotplot visualization.
}
\description{
Creates a dotplot visualization for enrichResult objects.
}
